+++ perf/WeihanLi.Npoi.Benchmark/Program.cs
@@ -1,0 +1,12 @@
+﻿using BenchmarkDotNet.Running;
+
+namespace WeihanLi.Npoi.Benchmark
+{
+    public class Program
+    {
+        public static void Main(string[] args)
+        {
+            BenchmarkRunner.Run<ExportExcelFileTest>();
+        }
+    }
+}